etl: write output files in batches
Adds --batch-size=xxx as a command line option. Defaults to 1 billion
right now, but that's subject to revision.

This also introduces a slight race condition with Athena queries.
We delete all files in the output directory before writing any
new ones (since batch counts may not line up and/or if we didn't,
as we overwrite one batch file, the contents could duplicate content
from an older batch file we haven't gotten to yet)

So there is a moment where Athena has incomplete data. Something to
fix in the future, but we know we'll want batching, so might as well
implement this now.