deid: add Scrubber class, handling more deid
This commit adds a bunch of refactoring and some new features to our
de-identification support. This is all done with an eye towards both
integration Microsoft's anonymizer tool as a next step and handling
many more FHIR resource types than the current five.

= Scrubbing =

The new Scrubber class is now the face of de-identification in
cumulus, and the Codebook class has moved to be an internal
implementation detail of the Scrubber.

Scrubber handles three de-identification steps:
1. ID substitution (via an internal Codebook instance)
2. Optionally stripping attachment data (used to be in etl.py)
3. Skipping resources with unrecognized modifier extensions,
   which arguably isn't really about de-identification, but it's
   a convienent place to do it.

While handling modifier extensions is the only truly new step
there, the whole process is now entirely generic and can be applied
to any new resource we want to scrub. For example, a deeply nested
reference or attachment will be scrubbed, without needing to write
custom FHIR resource code for it.

= ID Substitution =

The way ID substition works has also changed for non-patient,
non-encounter resources. Before, we'd generate entirely new UUIDs
for them, and only keep track of the ones we knew we'd need later
(namely, document references).

This won't scale as we add new resource types we care about.
So this commit instead will cryptographically hash incoming IDs to
generate outgoing IDs. While this is less secure than entirely random
IDs, it's a tradeoff to avoid keeping the entire mapping for all
resources in memory. And we use the same algorithm that Microsoft's
anonymization tool uses. Plus, the IDs don't even make it all the way
to the Cumulus dashboard anyway.

= Codebook DB Format =

The previous codebook format stored encounters underneath a patient.
Which meant that to look up the mapping for an encounter, you needed
a patient ID too. Once we start handling arbitrary resource types,
we may run into references to encounters that don't also include the
patient ID. And since encounter IDs are unique across all encounters,
it made sense to me to flatten that list in the codebook database.

So now all patients and all encounters are each just two separate
dictionaries. I also stopped storing encounter start end end periods,
since they were not being used.