Add a Dockerfile and wait for cTAKES on startup
This adds a toplevel Dockerfile to build a tiny little docker
image for cumulus. That will eventually help with deployments.

And also wait on the cTAKES port to be available on startup,
which also helps reduce some deployment scripting and prints a
nice error message for the user if a cTAKES server isn't running.