Support bulk exports from FHIR servers
This commit allows passing an http/https URL as the input path
when using the ndjson input format.

Cumulus ETL will then query the given server for bulk exports of
the resources it cares about.

Currently, we are very greedy and request everything the server
will give us. We can refine that in the future.

You'll also need to pass two new arguments when using this feature:
--smart-client-id: the client ID you registered (or path to one)
--smart-jwks: path to a file with the JWKS you registered