etl: allow running individual tasks to update single tables
This commit adds a new command line arg (--task) which can be
specified multiple times and/or include a list of comma-separated
task names.

Each task name is simply the output database name (i.e. symptom
for the symptom Athena table or patient, condition, etc).

In order to support this, we need to keep ID anonymization
consistent between separate task runs, because of cross-references.
So we now keep the hashing salt in the codebook.