oracle: actually let you connect against an Oracle server
We had some unused code we wrote to connect to an Oracle i2b2
server, but it was never hooked up to the main ETL loop.

This commit closes that gap. If you pass --input-format=i2b2 and
provide a tcp:// URL, we will assume an Oracle server and look
for the I2B2_SQL_USER and I2B2_SQL_PASS environment variables, to
connect.

- Adds the hash salt to the codebook
- Allows running one task at a time, via --task