oracle: actually let you connect against an Oracle server
We had some unused code we wrote to connect to an Oracle i2b2
server, but it was never hooked up to the main ETL loop.

This commit closes that gap. If you pass --input-format=i2b2 and
provide a tcp:// URL, we will assume an Oracle server and look
for the CUMULUS_SQL_USER and CUMULUS_SQL_PASSWORD environment
variables, to connect.