etl: read ndjson files written as standard bulk export names
The bulk-data-client (the node one, not our built-in one) [1] writes
files that look like 1.Condition.ndjson.

We were only recognizing files like Condition.ndjson or
Condition.1.ndjson (both styles that we use internally). But it makes
sense that a user might separately do a bulk export with the
bulk-data-client and then feed that directory to us. So this fixes
that oversight.

[1] https://github.com/smart-on-fhir/bulk-data-client