fix: merge schemas by default when writing to Delta Lake
Delta Lakes are super strict about schema validation when writing
to them. But FHIR is a bit unpredictable and a new batch of
a table might have fields that previous batch didn't (maybe the
hospital started adding metadata or stopped adding it)

So enable schema autoMerge when writing to a Delta Lake.

See https://delta.io/blog/2022-11-16-delta-lake-schema-enforcement/