fix: save the codebook periodically as we run the ETL
Previously, we'd save the codebook after every job. But jobs write
data out in batches as they go. So if the job is interrupted, that
means it has written out data that didn't make it to the codebook.

For patients and encounters, whose de-identified IDs are randomized
the first time they are encountered, this could cause a problem on
the next run, where the same patient is given a new ID because the
codebook was never saved from the previous run.

So this commit adds two changes:
- Save the codebook *before* writing out any batch of data.
- As an optimization, when the codebook wants to be saved, only write
  it out if anything has actually been modified inside of it. (Since
  we now want to save it more often, it made sense not to be saving
  a giant 1GB file all the time.)