feat: use ID hashing for Patient and Encounter too
We used ID hashing for all other types, and used a random UUID for
those types, which are treated specially because we wanted to
preserve their ID mappings for debugging purposes.

But this commit unifies the code paths and uses ID hashing for all
resource types. This gives the following benefits:

- We now write less often to the codebook. This will help us in
  the future when we are adding threading, because there is less
  coordination around shared writing that has to happen.
- We can save the cached mappings in a separate file that can be
  deleted, as some IRB requirements may have.
- We can add an option down the road that says "don't bother
  preserving mappings, I don't need debugging support"