feat(nlp): feed cTAKES a custom dictionary before an NLP task
This commit uses the new feature of our ctakes-covid docker to
accept custom dictionaries. We now provide it our covid dictionary
and wait for it to restart before proceeding to NLP.

The end result of this right now is no actual change in behavior.
The covid dictionary shipped with our default ctakes-covid docker
is the same as the one shipped with ctakesclient.

But (a) it's nice to have one source of truth for the dictionary
(in our case, ctakesclient) and (b) this will open the doors for
future studies with their own custom dictionaries.

An itemized list of interesting changes:
- Adjust our docker setup to share a volume between cTAKES and ETL.
  This is where the ETL can place override files (like a new
  dictionary).
- Moved ctakes.py into a new folder as nlp/extract.py (alongside
  the new file nlp/watcher.py)
- Adds a new argument --ctakes-overrides=DIR which defaults to
  /ctakes-overrides (the path used in our docker compose setup).
  If you are running ETL ourside of our compose setup for some
  reason, you will need this to run any NLP tasks.
- Before running an NLP task, we now feed cTAKES a new dictionary
  and wait for it to finish restarting.
- Bump the required ctakesclient version to 2.1 for some new API.
- Adjust our cTAKES tests to work against a real tiny http server
  rather than mocked out request calls. (This was required to
  emulate a remote socket dying on cTAKES restart.)