fix: use the correct base FHIR URL when downloading attachments
If you are doing a bulk export, the wrong base URL was being used
for attachments. It worked correctly for already-downloaded ndjson,
but not for a live export.

The fix is basically to preserve the input URL for BulkExporter,
but everywhere else use the root server URL.

In addition, this PR now always respects --fhir-url, as long as it
doesn't actively conflict with the input URL. This is more of a
defensive programming thing than a real fix. The above bug would
have had a workaround of passing --fhir-url if this behavior had
already been in place.