feat: disable philter by default, add --philter to enable
In early testing, philter is extremely slow. Like 10x to 240x slower.

So while we investigate how we can improve that, let's make it
optional for now. Most people hopefully don't *need* it, it was
designed as an extra safeguard. But folks can enable for now, opting
into the speed difference, while we look into this.