chore: handle new bandit warnings
The latest release of bandit (1.7.5) added a few new SQL checks
that we were tripping on.

It thought we might have SQL injection issues from doing string
formatting on raw SQL lines. But none of the cases it found were
based on user input, but rather were boilerplate generators.

I also moved some code into a test file because it was the only
place where the code was used.