feat: add chart-review subcommand
This is a new and still in-progress mode for cumulus-etl, designed to
assist in preparing chart reviews.

The idea is that you feed cumulus-etl some document references you want
to review (likely from de-identified Athena data), and it will
figure out which real documents map to those anonymous references,
run NLP & philter on the doc, then upload it to Label Studio.

Only the first part (downloading the real docs from anonymous IDs)
has been implemented in this commit.

The command is still hidden (i.e. does not show up on --help) and is
not documented. It doesn't really do anything except download and then
throw away some documents you asked for.

But this is a good checkpoint to build upon.