tests: work around python 3.10 async TestCase bug
Python 3.10 has this annoying bug where if you raise an unexpected
SystemExit, the test will hang. This caused us to go through all our
github CI minutes, and is annoying during development.

So as long as we still support Python 3.10, we can carry a small
wrapper TestCase implementation, to work around the issue.