feat: add new "convert" subcommand
This takes an ndjson output folder and converts it to a Delta Lake
output folder (or folds it into an existing one).

This is useful for workflows where manual review of the ndjson
output is required, but you do want to eventually get the data into a
Delta Lake.

Bare bones feature wise for now (and hopefully it can stay that way).

Extra refactoring:
- I moved some CLI subcommand logic down further into their modules.
- I noticed that one test wasn't being run (TestI2b2Loader) because it
  was an async test in an old-school (non-async) TestCase runner.
  To avoid similar accidents in the future, I switched all existing
  tests to subclass from our internal async runner.