fix: make sure we internally use valid S3 URLs after listing files
fsspec's S3 backend was giving us URLs without the s3:// prefix on
them, which just confused fsspec when we tried to use those URLs.
It thought they were local paths, reasonably.

So in Root(), we'll check for a backend doing that and prefix the
protocol.

This manifested as not being able to use an s3:// path as input
(you had to first download the files locally). This should now be
possible.