feat: add --errors-to=dir to capture errors for later analysis
This commit adds a new tools to your arsenal for debugging.

The new ETL command line argument --errors-to mirrors --export-to
but specifies a path to dump problem files. These are not in any
promised format, but the two places we use this new flag are:
* during upload to Delta Lake etc, if a batch fails, we store the
  batch in this folder -- it's anonymized by that point, so not
  100% useful, but it's something
* during NLP, we store the original docrefs for any failed extraction
  which will let you examine & retry

We also now only print one error message for a failure to download
text from a docref, not two.