fix: retry cTAKES when it times out
If cTAKES gets stuck in an unresponsive way, we'll now restart it
automatically -- this seems to unstick it. I'm guessing memory leak,
but we can always investigate that down the road.

ALso drop cTAKES timeout to 3m from 5m, given that we handle that
better, and both are overly generous anyway.