feat: strip HTML from all notes before passing to NLP
We found it can confuse NLP so bad that cTAKES will get stuck.

NLP results are slightly different with stripped notes, but it is a
minor change in covid symptom counts (<1%) and spot checking the
differences, it seems to just be around slightly different negation
conclusions -- which went both ways for html & stripped.

Or a few places where cTAKES gave up on long complicated HTML (so in
those cases, a clear improvement).

This commit adds stripping via the inscriptis library to the ETL and
changes the chart review mode to use inscriptis rather than html2text
(which gave decent results, but they weren't quite as good as
inscriptis and html2text is gpl-licensed, so we shouldn't ship that if
we want the resulting combo to be Apache licensed)