fix: drop fhirclient from dependency list, replace with our own auth
Ever since we stopped using fhirclient for parsing json, we've only
been using it for a tiny portion of our authentication flow.

So tiny that it doesn't make sense to keep the dependency around,
especially since the critical piece of that auth flow we use was only
in a PR branch that hasn't been merged yet.

This drops:
- fhirclient
- responses (for tests)
- urllib3 (for tests -- this could have been dropped a while back)

Replacing the fhirclient 3-call flow:
- FhirClient()
- prepare()
- authenticate()

We now have a single POST call from which we grab the auth token
ourselves. We don't need a lot of the complexity of oauth2 (like
refresh tokens or non-JWT auth flows).