fix: avoid schema-mismatch errors for int columns that have nulls
Pandas converts integer columns with nulls to float columns by
default. This confused our new schema merging feature.

So now we ask Pandas to use more modern column types (ones that can
handle nulls better). This fixes the schema issue.

BUT it might introduce incompatibilities with existing tables.
If so, the easiest fix is to blow away the table and recreate it.