fix: avoid partial writes of files through fsspec
This introduces a further fix to atomic_open() which guarantees that
we close the file when done with it (thus flushing all remaining
bytes out to disk).

It also fixes up the S3 mocking support to be more robust and avoid
some of the heavy patching we were doing for aiobotocore's sake.