feat(deltalake): don't overwrite newer rows if we can avoid it
If both the lake table and the update rows have meta.lastUpdated,
only push the update row if it is actually a newer version.

This should let folks with meta.lastUpdated support more safely mix
and match batches of ndjson without having to stress about when
each batch was exported from the EHR.

Unfortunately Epic doesn't provide this field, so this benefit is
limited to other EHRs.