feat: order tasks to minimize codebook writes
Prior to this commit, tasks were either:
A) In alphabetical order if no task list was provided
B) In arbitrary order if a task list was provided (the task list was
   turned into a set at some point in the processing)

**No task list**

Let's address (A) first - what order should we run them if no task list
is given? While alphabetical is fine, we have an opportunity to reduce
writes to the cached codebook mapping file.

Every time we come across a new encounter or patient ID, we first write
out the mapping file before writing the batch to the output target.

To reduce churn there, we can simply run the tasks with the highest
density of new IDs first and have fewer writes of the mapping file.

So in the case of no provided task order, we now run 'encounter' first,
then 'patient', then the rest in alphabetical order.

**User task list**

Obviously, an arbitrary order makes no sense. So either we could let
the user or the ETL have control over ordering when a task list is
provided.

But since the ETL does have opinions here (see above), I figure might
as well keep control over the order. We can always reverse this
thinking later, but once we give the user control over the order, it's
harder to wrest it back.

So when the user gives a list, we'll run those tasks, but in the order
the ETL wants (the same order as when no task list is provided).