feat: add cross-task temporary dir to share downloaded notes
This commit adds a global temporary directory that can be used to share
data between disparate bits of code/tasks.

Specifically, the first user of this is fhir_utils.get_docref_note
which now caches any downloaded notes in this temporary directory.

This will make it much more performant to add a second NLP task (it
won't have to re-download the notes).

Like all temporary directories managed via context manager, this
directory will be cleaned up upon ETL finishing or being interrupted
by an error.

It is also created with mode 0700, so only the ETL user can read the
files.