feat!: correctly delete old symptoms if NLP now returns zero
The goal of this commit is to handle the following case:
- Yesterday you ran NLP and got five symptoms.
- Today you run NLP and get zero symptoms.

Before this commit, the previous five symptoms would remain in Athena.
With this commit, they are all deleted from Athena.

This may not sound super common, but it does happen if you change the
NLP pipeline (like during the development of an NLP task) or the note
text changes.

The ETL currently handles *reduction* of symptoms by defining a group
field of `docref_id` and any symptoms that match that group but aren't
in the current batch get deleted. But this previously relied on at
least one row being present for any group to get noticed.

So if you have zero rows, your group would be left alone, even though
NLP says all the group members should be deleted.

Changes in this commit:
- Tasks that use this group feature now must also keep track of the
  group values as it generates rows.
- Because the `convert` command would not normally see these group
  definitions (because there is no output ndjson row for them), we now
  write out a metadata file for each ndjson batch. So for a file like
  `table.000.ndjson`, there can now be a sibling file `table.000.meta`
  which holds batch metadata, currently just this groups definition.
  The `convert` command now passes that along to deltalake.
- Rather than also fix the parquet output format to write out this
  metadata, I just removed the format. I didn't _need_ to do this, but
  I figured we don't need this format anymore. If you don't care about
  human readability, you should just be using the deltalake format.
  This format has outlived its purpose, so here's a bit of cleanup.
- Now the only remaining output formats are ndjson and deltalake.
- Because I added this breaking change (which should affect no one),
  I finally bumped our version to 1.0.0 -- this reflects not just this
  change but the general production-readiness of Cumulus ETL. We still
  aren't doing actual releases, so semantic versioning is not 100%
  useful here, but it's just a possible indication to close-watchers
  that something big has changed.