feat: add generated_on timestamp to covid NLP results
This is a little late to be super useful for the covid study, but as a
best-practices change that might inform future copy/pasted NLP tasks,
it is good to have a timestamp.

This is not all the information we'd like to have about the generation
of the NLP results - ideally lots of info about the source doc
(docRefLastUpdated or docRefChecksum?) and NLP algorithm (what commit
of cTAKES and what version of our custom covid dictionary? What about
cNLP? etc).

But this seems like bare minimum metadata that should have been there
from the start.