feat: skip MS tool de-id for solo NLP task runs
If an ETL run is only for an NLP task, the MS tool de-identification
is unnecessary. The NLP task only pulls out certain bits of the
incoming resources and still does the standard ID scrubbing itself.

So it's just wasted time, especially since NLP tasks are frequently
run by themselves as a separate GPU run.