feat: use latest ctakesclient (with updated icd10 codes)
The new ctakesclient modified the covid icd10 list, and this commit
pulls that in.

This also adds a new field to the covid symptom output, "task_version"
which will get incremented every time the NLP algorithm changes (like
the icd10 code change might).

The change for this would be very small (maybe slightly more symptoms
caught). But it's nice to have the mechanism to talk about task
versioning.