feat(covid-symptom): add BCH/Cerner-specific ED codes
Previously, we had some BCH/i2b2-specific ED codes for BCH's i2b2 data.
And the LOINC codes were supposed to be sufficient for the properly
coded data coming from Cerner.

But it turns out that not all DocRefs in Cerner have proper LOINC
codes, so this commit adds the same site-specific i2b2 codes but with
a Cerner "system".

This should be harmless, study-wise since all these NLP symptoms are
filtered down on the Library side - worst case is that we run NLP on
docs we don't need to. But this now allows the Library side to expand
what it looks for, if it wants.