fix: guard against accidental deletes with ndjson output format
The ndjson output formatter deletes all existing content in its target
directory, because it is not an incremental formatter.

But this introduces the possibility of accidentally using the ndjson
output format on a delta lake folder or some other important folder.

So this commit checks the contents of the target output dir and errors
out if there are odd looking files in there.