feat(chart-review): push internal docref spans to Label Studio
In case we need to know the internal placing of doc notes,
when we gather multiple notes into one note, push up the start &
end of those texts.

No planned use for this right now, but I want to get this in so
when we do want it, the metadata already exists.