fix: write "null match" NLP rows for DocRefs with no symptoms
Previously, there was no way to detect the difference between a DocRef
that was not processed (not an ED DocRef or an error occurred etc) vs
a DocRef that had no symptoms.

Both just had zero rows in the Athena db.

This commit starts adding a single row in the nlp_results table with
a null "match" column for this case.

Existing SQL should be able to gracefully handle this change.
If/when we want to update the SQL to handle these kinds of DocRefs,
it can be updated to treat these new rows specially. But for now,
they will be ignored.