feat: downgrade some fatal connection errors to warnings
Previously, if you were running an NLP task and we needed to download
a clinical note but couldn't, we'd early-exit with an error and say
"please provide the auth flags, kthx"

Now, we merely warn about it once but keep processing DocRefs.

This is designed to more gracefully handle some cases like BCH has,
where their DocRefs contain some URLs that always error out, but the
other DocRefs have been nicely inlined.

So we want to be able to run in a "no-network" context where we don't
need to talk to the FHIR server at all. We just want to process the
already-inlined notes.

In other contexts, where the user genuinely made a mistake and should
re-run with auth parameters, this should still be plenty of notice.
There is an error printed to the console, and the job is marked as
failing, so the ETL as a whole will fail with a status code and
error message at the end.