feat(upload-notes): add ability to tag philter spans, not redact them
This commit deprecates --no-philter in favor of --philter=disable
and adds --philter=label (and the default --philter=redact).

When --philter=label is requested, we send a prediction layer to Label
Studio that highlights all the detected PHI spans for you.

This is helpful when doing manual de-identification, since philter can
do some of the grunt work for you.