bulk: store "deleted" and "error" results in export subfolders
If the bulk export server provides an array of error messages or
an array of deleted resources, make sure we persist those on disk.

Previously, we only printed the error messages to the console.
We still do that, but also write them to an error/ subfolder.

Previously, we ignored any deleted resources.
Now we write them to a deleted/ subfolder.

This matches the behavior of bulk-data-client.