feat: mark data/url fields inside DocRef attachments as absent
When we strip the data & url fields from DocumentReferences, we
previously just deleted the fields and moved on.

But in order to be able to do some QA analysis on whether those fields
were provided in the first place, it's nice to have a record of them.

So whenever we delete those fields, we also now leave a data-absent
extension in place with the value "masked".