Always prefer resource rows that come later
When iterating through the input ndjson, if we see the same row ID
inside a batch, we now prefer the row that came later.

This means that our behavior is more consistent:
- If rows are in same batch, later is preferred (new)
- If rows are across batches, later is preferred (same)

This way, you can safely throw ndjson updates to your original FHIR
data alongside the original data and it can reliably be processed
(as long as you choose good filename ordering).