fix: improve FHIR-root-URL parsing to cover more cases
Before, we were just using some regexes on the URL, but let's
actually use a parser and separate out the pieces.

This is better in general, but also happens to let us handle cases
like the user putting an export URL like ".../$export?_type=..." on
the shell, having the shell zero-out the $export string because it
thinks it is a variable, and then us seeing a URL like "/?_type=..."
(which isn't really a valid URL, but we will insert missing $export
strings later in the process).

As part of this, I refactored some instances of URL parsing into one
single FhirUrl class, which we can build upon in the future as needed.