fix(upload-notes): always send label list, even with static labels
It turns out that Label Studio can nowadays always accept a label
list, it doesn't need to only be sent when dynamic labels are
turned on. And always sending it helps avoid a situation when we
don't think we need to send it but we do (like when there are both
static and dynamic labels in play).