feat(export): add a --cancel option
This helps recover from situations where the server only lets you make
so many exports at once, and you don't need a previous one.

Or just to be a good citizen.

Note that the ETL normally cancels/cleans up for you in the success
case. But if the ETL gets interrupted, this lets you manually clean up.