feat: send SMART v2 scopes if the server supports them
Most servers should (per spec: SHOULD) still support v1 scopes,
even if they also support v2. But since it's only a recommendation,
let's prepare ourselves for a future server that drops v1 support.

We don't currently do any parsing of the returned actually-approved
scopes (we just assume it all worked out and we'll error out when we
try to do an export), so this commit only needs to worry about the
scopes that we send to the server.