fix(export): use a consistent exportId across export restarts
Previously, we followed bulk-export-client's lead and generated a new
uuid for each run of the client. Which in the case of resuming a
previous interrupted export, means that the same "logical" export has
two different IDs.

Which, could be fine. But we also separately try to parse info from
the log like export group name and export timestamp. It's a lot easier
to do that with consistent export IDs.

We maybe should also investigate making that log reading robust
against clients that don't set consistent IDs, but this is an easy
change for now to make our client a little more consistent.

(The spec does not give any required form to the ID, so as far as I
can tell, it's an opaque string that we can stuff a URL into.)