feat(upload-notes): disable NLP by default
It feels surprising to folks that they need cTAKES to upload notes.
And at BCH, we typically disable it all the time anyway. Let's
go with the simple path by default, but allow it to be enabled.

The old --no-nlp arg will persist, but do nothing. The new --nlp
arg will turn on NLP.