feat(deid): allow most Epic systems back again, detect custom codes
If an internal Epic system has a text (non-numeric) code we strip
the code and display. If it has a code of 0, we strip the display.

This should catch the examples of freeform custom codes & displays
we've seen in the wild, while leaving the other "standard" Epic
codes in place.