feat(upload-notes): add --grouping={encounter,none} flag
This lets you turn off grouping notes by encounter, if that is
not helpful for the kind of chart review you're doing.

Note that we keep using the "enc_id" field in Label Studio as the
unique task ID field, to avoid breaking uploading to existing
projects. We just push a ref like DocumentReference/A to that field
in the case we aren't grouping by encounter. (just like we do when
grouping by encounter but there isn't an encounter!)