fix: error out if we detect a new PHI folder for same output folder
This adds a new field to the codebook (`codebook_id`) that gets set
on codebook load (and backfilled if not present). This lets us
identify a PHI folder/codebook in a safe way.

Then each JobConfig writes that codebook ID out, and we confirm the
most recent JobConfig and the current PHI folder match.