Convert our previous hugging face support from TGI to vLLM
vLLM seems like the future and lets us consolidate on one client
API (openai).

So this converts the "hftest__summary" task to use a vLLM image for
Llama2.

Note that hftest__summary is a toy/example of how we'd integrate a
model into a test. It's just meant as a base for future tasks.
So this commit just improves our example.