feat!: add --select-by-* options to upload-notes, remove downloading
This adds the --select-by-word, -regex, -athena-table, -csv, and
-anon-csv options to upload-notes, deprecating the previous --docrefs
and --anon-docrefs options (which now map to by-csv and by-anon-csv).

This also removes the rarely used "download notes from your EHR on
the fly" feature which meant you didn't need local copies of the notes.
It could even do a bulk export if you didn't provide a --docrefs flag.
It is too hard to square with the new set of options and is a lot of
complexity for a feature we don't really use and don't really recommend
to sites to use (you should be archiving your notes for easy access
anyway).

Downloading the wrapped clinical notes referred to by your NDJSON is
still supported! This is just removing the ability to go direct to your
EHR for the DocRefs and DxReports themselves.