nlp: improve bedrock support
Now we should actually work for multiple models, using Bedrock's own
API, not the OpenAI API (which only works for OpenAI models).

This also adds --provider as an argument for `cumulus-etl nlp`, which
specifies which NLP provider to use (azure, bedrock, or local).
The default is local, and if you try to use the wrong model, we now
warn about it in the console (and I added a model support table to the
docs).

Previously we were using environment variables to signal which to use,
but now there won't be an obvious difference between bedrock and local,
since bedrock just uses your typical AWS config, rather than custom env
vars.